+++ ft_memset.c
@@ -6,24 +6,24 @@
 /*   By: alidy <alidy@student.le-101.fr>            +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2019/10/08 15:17:24 by alidy        #+#   ##    ##    #+#       */
-/*   Updated: 2019/10/09 15:00:03 by alidy       ###    #+. /#+    ###.fr     */
+/*   Updated: 2019/10/10 10:59:48 by alidy       ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
 
-#include <string.h>
+#include "libft.h"
 
-void        *ft_memset(void *b, int c, size_t len)
+void	*ft_memset(void *b, int c, size_t len)
 {
-    int i;
-    unsigned char *b1;
+	size_t			i;
+	unsigned char	*b1;
 
-    i = 0;
-    b1 = b;
-    while (i < len)
-    {
-        b1[i] = c;
-        i++;
-    }
-    return (b);
+	i = 0;
+	b1 = b;
+	while (i < len)
+	{
+		b1[i] = c;
+		i++;
+	}
+	return (b);
 }
