+++ ft_substr.c
@@ -6,14 +6,16 @@
 /*   By: alidy <alidy@student.le-101.fr>            +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2019/10/09 10:15:14 by alidy        #+#   ##    ##    #+#       */
-/*   Updated: 2019/10/09 10:44:14 by alidy       ###    #+. /#+    ###.fr     */
+/*   Updated: 2019/10/10 11:02:55 by alidy       ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
 
+#include "libft.h"
+
 char	*ft_substr(char const *s, unsigned int start, size_t len)
 {
-	int 	size;
+	size_t	size;
 	char	*s1;
 
 	size = 0;
